Add type/provider pairs for federation
This allows us to manage federation upstreams, sets of federation
upstreams and policies (which are used to setup federation and also
other things inside rabbitmq).

TODO: No test cases!

Conflicts:
	lib/puppet/provider/rabbitmq_exchange/rabbitmqadmin.rb

FIx dependency ordering

Fix HOME to be the right place for where we put the rabbitmqadmin credentials

Spaces are explicitly allowed

Remove all type checking as may be an array